added a meshkit error
